move refactoring pytest files
