+++ tests/resources/FOSSLight-Report_sample.xlsx
