+++ tests/test_spdx_licenses.py
@@ -4,9 +4,9 @@ from fosslight_util.spdx_licenses import get_spdx_licenses_json
 
 
 def test_get_spdx_licenses_json():
-    #when
+    # when
     success, _, licenses = get_spdx_licenses_json()
 
-    #then
+    # then
     assert success is True
     assert len(licenses) > 0
