+++ tox.ini
@@ -8,6 +8,7 @@ skipdist = true
 install_command = pip install {opts} {packages}
 allowlist_externals = cat
                       ls
+                      pytest
 setenv =
   PYTHONPATH=.
 
@@ -17,65 +18,27 @@ exclude = .tox/*
 
 [pytest]
 filterwarnings = ignore::DeprecationWarning
-norecursedirs = test_result/*
+norecursedirs = test_result/* tests/legacy
 
 [testenv:test_run]
 deps =
   -r{toxinidir}/requirements-dev.txt
 wheel = true
+# move to tests dir
+changedir = {tox_root}/tests
 commands =
-  # Test - logging
-  python tests/test_log.py
-  ls test_result
-  cat test_result/log_file1.txt
-  # Test - writing excel and csv
-  python tests/test_excel_and_csv.py
-  ls test_result/excel_and_csv/
-  cat test_result/excel_and_csv/excel/Test_Excel.xlsx
-  cat test_result/excel_and_csv/csv/Test_Csv.csv
-  # Test - downloading source
-  fosslight_download  -s "https://github.com/LGE-OSS/example" -t test_git/
-  ls test_git/
-  fosslight_download -s "https://pypi.org/project/fosslight-dependency/3.0.5/" -t test_wget -d test_logs
-  fosslight_download -s "https://www.npmjs.com/package/json-schema/v/0.3.0" -t test_wget -d test_logs
-  fosslight_download -s "https://mvnrepository.com/artifact/com.fasterxml.jackson.core/jackson-databind/2.12.2" -t test_wget -d test_logs
-  fosslight_download -s "https://pub.dev/packages/file/versions/5.2.1" -t test_wget -d test_logs
-  ls test_wget/
-  # Test - write output file
-  python tests/test_output_format.py
-  # Test - get spdx licenses
-  python tests/test_spdx_licenses.py
+# Test - run pytest
+    pytest
 
 [testenv:release]
 deps =
   -r{toxinidir}/requirements-dev.txt
 wheel = true
+# move to tests dir
+changedir = {tox_root}/tests
 commands =
-  # Test - print help msg
-  python tests/test_help.py
-  # Test - logging
-  python tests/test_log.py
-  ls test_result
-  cat test_result/log_file1.txt
-  # Test - writing excel
-  python tests/test_excel_and_csv.py
-  ls test_result/excel_and_csv/
-  cat test_result/excel_and_csv/excel/Test_Excel.xlsx
-  cat test_result/excel_and_csv/csv/Test_Csv.csv
-  # Test - writing opossum
-  python tests/test_opossum.py
-  ls test_result/opossum
-  # Test - writing yaml
-  python tests/test_yaml.py
-  cat test_result/yaml/FL-TEST_yaml.yaml
-  # Test - timer
-  python tests/test_timer.py
-  # Test - downloading source
-  python tests/test_download.py
-  ls test_result/download
-  # Test - write output file
-  python tests/test_output_format.py
-  # Test - get spdx licenses
-  python tests/test_spdx_licenses.py
+  # Test - run pytest
+  pytest
+
   # Test - check PEP8
   pytest -v --flake8
